Reorganizing files to dir structure. Adding events route for admin
