+++ server/Managers/Events/AdminEvents.py
@@ -1,0 +1,6 @@
+
+
+
+
+class AdminEventsManager:
+    pass
