+++ server/Routes/Auth/login.py
@@ -2,7 +2,7 @@ from http import HTTPStatus
 from flask import make_response, request, abort, jsonify
 from flask_restful import Resource
 
-from server.Managers.UserManager import UserManager
+from server.Managers.Auth.UserManager import UserManager
 
 class LoginRoute(Resource):
     def post(self):
